some more feat event updates

+++ events/FEAT_Events/metadlg.py
@@ -35,11 +35,10 @@
 |       Nathan Lauffenburger   <nathan.lauffenburger@noaa.gov>
 """
 
-from PyQt6.QtCore import *
-from PyQt6.QtGui import *
 from PyQt6.QtWidgets import *
 from ui import ui_MetaDlg
 import numpad
+import keypad
 import messagedlg
 
 
@@ -60,34 +59,36 @@ class MetaDlg(QDialog, ui_MetaDlg.Ui_metaDlg):
         self.errorIcons = parent.errorIcons
         self.gear = parent.gear
         self.scientist = parent.scientist
-        self.event_entered = False
+        self.event_entered = parent.event_entered
 
         self.message = messagedlg.MessageDlg(self)
 
+        # this is all hard-coded for now todo: should use DB for this later
         self.required_params = ['Transect', 'TrawlScientist', 'Gear']
 
-        self.pbs = {'Transect': self.pb_transect,
-                    'TargetDepth': self.pb_td,
-                    'TDLatitude': self.pb_tdlat,
-                    'TDLongitude': self.pb_tdlon,
-                    'WindSpd': self.pb_wind_spd,
-                    'WindDir': self.pb_wind_dir,
-                    'SwellHeight': self.pb_swell_h,
-                    'WindWaves': self.pb_wind_waves}
-        self.cbs = {'TrawlScientist': self.cb_sci,
-                    'Gear': self.cb_gear,
-                    'TomWeights': self.cb_toms,
-                    'HeadropeSBE': self.cb_head_sbe,
-                    'FootropeSBE': self.cb_foot_sbe,
-                    'CameraSBE': self.cb_cam_sbe,
-                    'CameraType': self.cb_cam_type,
-                    'CameraView': self.cb_cam_view}
+        self.pbs = {'Transect': [self.pb_transect, 'ed', 'np'],
+                    'TargetDepth': [self.pb_td, 'ed', 'np'],
+                    'TDLatitude': [self.pb_tdlat, 'ed', 'kp'],
+                    'TDLongitude': [self.pb_tdlon, 'ed', 'np'],
+                    'WindSpd': [self.pb_wind_spd, 'ed', 'np'],
+                    'WindDir': [self.pb_wind_dir, 'ed', 'np'],
+                    'SwellHeight': [self.pb_swell_h, 'ed', 'np'],
+                    'WindWaves': [self.pb_wind_waves, 'ed', 'np']}
+        self.cbs = {'TrawlScientist': [self.cb_sci, 'ed'],
+                    'Gear': [self.cb_gear, 'ed'],
+                    'TomWeights': [self.cb_toms, 'ga'],
+                    'HeadropeSBE': [self.cb_head_sbe, 'ga'],
+                    'FootropeSBE': [self.cb_foot_sbe, 'ga'],
+                    'CameraSBE': [self.cb_cam_sbe, 'ga'],
+                    'CameraType': [self.cb_cam_type, 'ga'],
+                    'CameraView': [self.cb_cam_view, 'ga']}
 
         # load dropdown boxes
         self.load_dropdowns()
 
-        # set up numpad
+        # set up numpad ane keypad
         self.numpad = numpad.NumPad(self)
+        self.keypad = keypad.KeyPad('', self)
 
         # deal with it if it is a reload of data
         if self.reloaded:
@@ -96,33 +97,36 @@ class MetaDlg(QDialog, ui_MetaDlg.Ui_metaDlg):
 
         # set signals and slots
         for param, btn in self.pbs.items():
-            btn.clicked.connect(self.enter_data)
+            btn[0].clicked.connect(self.enter_data)
         self.pb_save.clicked.connect(self.save)
         self.pb_cancel.clicked.connect(self.cancel)
 
     def reload_data(self):
         """
-        populates the widgets with entered data
+        populates the widgets with data loaded from the database
         :return: none
         """
         # refill any buttons
-        for param, pb in self.pbs.items():
+        for param, pb_lst in self.pbs.items():
+            pb, table, np = pb_lst
             # see if it exists
-            exists = self.check_if_exists(param)
+            exists = self.check_if_exists(param, table)
             # set text if so
             if exists != 0:
                 pb.setText(exists)
         # refill any drop downs
-        for param, cb in self.cbs.items():
+        for param, cb_lst in self.cbs.items():
+            cb, table = cb_lst
             # see if it exists
-            exists = self.check_if_exists(param)
+            exists = self.check_if_exists(param, table)
             # set text if so
             if exists != 0:
                 cb.setCurrentText(exists)
 
     def load_dropdowns(self):
         """
-
+        loads the dropdowns from the gear_accessory_option table
+        todo: hard-coded for now, but should be updated when a fix is available
         :return:
         """
         # list of scientists for the fisher
@@ -181,22 +185,36 @@ class MetaDlg(QDialog, ui_MetaDlg.Ui_metaDlg):
         allows user to enter the value using a number pad and saves that value as the button text
         :return: none
         """
-        # send up numpad
-        self.numpad.msgLabel.setText("Enter value")
-        if not self.numpad.exec():
-            return
+        btn_name = self.sender().objectName()
+        # get the dialog to send up (numpad or keypad)
+        np_name = ''
+        for txt, lst in self.pbs.items():
+            btn, table, np = lst
+            if btn_name == btn.objectName():
+                np_name = np
+
+        if np_name == 'np':
+            self.numpad.msgLabel.setText("Enter value")
+            if not self.numpad.exec():
+                return
+            val = self.numpad.value
+        else:
+            if not self.keypad.exec():
+                return
+            val = self.keypad.dispEdit.toPlainText()
         # save value as text of button
-        self.sender().setText(self.numpad.value)
+        self.sender().setText(val)
 
     def write_haul_record(self):
         """
-        writes the initial records into the events table
+        writes the initial records into the events table;
+        todo: default for event type is used, but could be taken from db at some point
         :return:
         """
         #  write record to events table
         sql = ("INSERT INTO " + self.schema + ".events (ship, survey, event_id, gear, event_type, " +
                 "performance_code, scientist, comments) VALUES (" + self.ship + "," + self.survey + "," +
-               str(self.activeEvent) + ",'" + self.gear + "', 0, 0, '" + self.scientist + "', '')")
+               str(self.activeEvent) + ",'" + self.gear + "', 8, 0, '" + self.scientist + "', '')")
         self.db.dbExec(sql)
         # set the flag to true that the event was entered
         self.event_entered = True
@@ -215,47 +233,70 @@ class MetaDlg(QDialog, ui_MetaDlg.Ui_metaDlg):
         required = False
         for r in self.required_params:
             try:
-                p_meas = self.pbs[r].text()
-            except:
-                p_meas = self.cbs[r].currentText()
+                p_meas = self.pbs[r][0].text()
+            except KeyError:
+                p_meas = self.cbs[r][0].currentText()
             if p_meas == '':
                 required = True
         if required:
             self.message.setMessage(self.errorIcons[0], self.errorSounds[0],
                                     "You need to add all of the required parameters before you can continue!",
                                     'info')
-            self.message.exec()
-            # todo: dont close the metadata window after ok
+            if self.message.exec():
+                self.reload_data()
+                # this is the only way I could find that would stop the meta dialog from being closed when this
+                # dialog was closed - fix if there is a better way
+                return 'stub'
         else:
             # iterate over all buttons and dropdowns
-            for param, pb in self.pbs.items():
-                # check if it already exists in EVENT_DATA
-                exists = self.check_if_exists(param)
+            for param, pb_lst in self.pbs.items():
+                pb, table, np = pb_lst
+                # check if it already exists in the table
+                exists = self.check_if_exists(param, table)
                 # if the entry doesn't exist and the text is not blank, enter a new row
                 if exists == 0 and pb.text() != '':
-                    le_sql = ("INSERT INTO " + self.schema +
-                              ".event_data (ship, survey, event_id, partition, event_parameter, "
-                              "parameter_value) VALUES (" + self.ship + ", " + self.survey + ", "
-                              + str(self.activeEvent) + ", 'MainTrawl', '" + param + "', '" + pb.text() + "')")
+                    if table == 'ed':
+                        le_sql = ("INSERT INTO " + self.schema +
+                                  ".event_data (ship, survey, event_id, partition, event_parameter, "
+                                  "parameter_value) VALUES (" + self.ship + ", " + self.survey + ", "
+                                  + str(self.activeEvent) + ", 'MainTrawl', '" + param + "', '" + pb.text() + "')")
+                    else:
+                        le_sql = ("INSERT INTO " + self.schema +
+                                  ".gear_accessory (ship, survey, event_id, gear_accessory, gear_accessory_option) "
+                                  "VALUES (" + self.ship + ", " + self.survey + ", "
+                                  + str(self.activeEvent) + ", '" + param + "', '" + pb.text() + "')")
                     self.db.dbQuery(le_sql)
                 # if the entry exists and the text is not blank, update with the new text
                 elif exists != 0 and pb.text() != '':
-                    le_sql = ("UPDATE " + self.schema +
-                              ".event_data SET parameter_value = '" + pb.text() + "' WHERE ship=" + self.ship +
-                              " AND survey=" + self.survey + " AND event_id=" + str(self.activeEvent) +
-                              " AND partition='MainTrawl' AND event_parameter='" + param + "'")
+                    if table == 'ed':
+                        le_sql = ("UPDATE " + self.schema +
+                                  ".event_data SET parameter_value = '" + pb.text() + "' WHERE ship=" + self.ship +
+                                  " AND survey=" + self.survey + " AND event_id=" + str(self.activeEvent) +
+                                  " AND partition='MainTrawl' AND event_parameter='" + param + "'")
+                    else:
+                        le_sql = ("UPDATE " + self.schema +
+                                  ".gear_accessory SET gear_accessory_option = '" + pb.text()
+                                  + "' WHERE ship=" + self.ship + " AND survey=" + self.survey + " AND event_id="
+                                  + str(self.activeEvent) + " AND gear_accessory='" + param + "'")
                     self.db.dbQuery(le_sql)
                 # if the entry exists and the text is blank, delete the row
                 elif exists != 0 and pb.text() == '':
-                    le_sql = ("DELETE FROM " + self.schema +
-                              ".event_data WHERE ship=" + self.ship +
-                              " AND survey=" + self.survey + " AND event_id=" + str(self.activeEvent) +
-                              " AND partition='MainTrawl' AND event_parameter='" + param + "'")
+                    if table == 'ed':
+                        le_sql = ("DELETE FROM " + self.schema +
+                                  ".event_data WHERE ship=" + self.ship +
+                                  " AND survey=" + self.survey + " AND event_id=" + str(self.activeEvent) +
+                                  " AND partition='MainTrawl' AND event_parameter='" + param + "'")
+                    else:
+                        le_sql = ("DELETE FROM " + self.schema +
+                                  ".gear_accessory WHERE ship=" + self.ship +
+                                  " AND survey=" + self.survey + " AND event_id=" + str(self.activeEvent) +
+                                  " AND gear_accessory='" + pb.text() + "'")
                     self.db.dbQuery(le_sql)
-            for param, cb in self.cbs.items():
-                # check if it already exists in EVENT_DATA
-                exists = self.check_if_exists(param)
-                # if it is for the trawl scientist or the gear, update the EVENTS table
+            for param, cb_lst in self.cbs.items():
+                cb, table = cb_lst
+                # check if it already exists
+                exists = self.check_if_exists(param, table)
+                # if it is for the trawl scientist or the gear, update the EVENTS table as well
                 if param == 'TrawlScientist':
                     update_sci_sql = ("UPDATE " + self.schema +
                                       ".events SET scientist='" + cb.currentText() + "' WHERE ship="
@@ -270,36 +311,62 @@ class MetaDlg(QDialog, ui_MetaDlg.Ui_metaDlg):
                     self.db.dbQuery(update_gear_sql)
                 # if it doesn't exist and the text is not blank, add a new row
                 if exists == 0 and cb.currentText() != '':
-                    cb_sql = ("INSERT INTO " + self.schema +
-                              ".event_data (ship, survey, event_id, partition, event_parameter, "
-                              "parameter_value) VALUES (" + self.ship + ", " + self.survey + ", "
-                              + str(self.activeEvent) + ", 'MainTrawl', '" + param + "', '" + cb.currentText() + "')")
+                    if table == 'ed':
+                        cb_sql = ("INSERT INTO " + self.schema +
+                                  ".event_data (ship, survey, event_id, partition, event_parameter, "
+                                  "parameter_value) VALUES (" + self.ship + ", " + self.survey + ", "
+                                  + str(self.activeEvent) + ", 'MainTrawl', '" + param + "', '"
+                                  + cb.currentText() + "')")
+                    else:
+                        cb_sql = ("INSERT INTO " + self.schema +
+                                  ".gear_accessory (ship, survey, event_id, gear_accessory, gear_accessory_option) "
+                                  "VALUES (" + self.ship + ", " + self.survey + ", " + str(self.activeEvent)
+                                  + ", '" + param + "', '" + cb.currentText() + "')")
                     self.db.dbQuery(cb_sql)
                 # if it exists and the text is not blank, update the row
                 elif exists != 0 and cb.currentText() != '':
-                    cb_sql = ("UPDATE " + self.schema +
-                              ".event_data SET parameter_value = '" + cb.currentText() + "' WHERE ship="
-                              + self.ship + " AND survey=" + self.survey + " AND event_id=" + str(self.activeEvent) +
-                              " AND partition='MainTrawl' AND event_parameter='" + param + "'")
+                    if table == 'ed':
+                        cb_sql = ("UPDATE " + self.schema +
+                                  ".event_data SET parameter_value = '" + cb.currentText() + "' WHERE ship="
+                                  + self.ship + " AND survey=" + self.survey + " AND event_id="
+                                  + str(self.activeEvent) + " AND partition='MainTrawl' AND event_parameter='"
+                                  + param + "'")
+                    else:
+                        cb_sql = ("UPDATE " + self.schema +
+                                  ".gear_accessory SET gear_accessory_option = '" + cb.currentText() + "' WHERE ship="
+                                  + self.ship + " AND survey=" + self.survey + " AND event_id="
+                                  + str(self.activeEvent) + " AND gear_accessory='" + param + "'")
                     self.db.dbQuery(cb_sql)
                 # if it exists and the text is blank, delete the row
                 elif exists != 0 and cb.currentText() == '':
-                    cb_sql = ("DELETE FROM " + self.schema +
-                              ".event_data WHERE ship="
-                              + self.ship + " AND survey=" + self.survey + " AND event_id=" + str(self.activeEvent) +
-                              " AND partition='MainTrawl' AND event_parameter='" + param + "'")
+                    if table == 'ed':
+                        cb_sql = ("DELETE FROM " + self.schema +
+                                  ".event_data WHERE ship="
+                                  + self.ship + " AND survey=" + self.survey + " AND event_id="
+                                  + str(self.activeEvent) + " AND partition='MainTrawl' AND event_parameter='"
+                                  + param + "'")
+                    else:
+                        cb_sql = ("DELETE FROM " + self.schema + ".gear_accessory WHERE ship="
+                                  + self.ship + " AND survey=" + self.survey + " AND event_id="
+                                  + str(self.activeEvent) + " AND gear_accessory='" + param + "'")
                     self.db.dbQuery(cb_sql)
         self.accept()
 
-    def check_if_exists(self, event_parameter):
+    def check_if_exists(self, event_parameter, table):
         """
         checks if the entry exists in the database
         :param event_parameter: the parameter to check
+        :param table: the table to query
         :return: the value of the parameter if exists and 0 if not
         """
-        exists_sql = ("SELECT parameter_value FROM " + self.schema + ".event_data WHERE ship=" + self.ship +
-                      " AND survey=" + self.survey + " AND event_id=" + str(self.activeEvent) +
-                      " AND partition='MainTrawl' AND event_parameter='" + event_parameter + "'")
+        if table == 'ed':
+            exists_sql = ("SELECT parameter_value FROM " + self.schema + ".event_data WHERE ship=" + self.ship +
+                          " AND survey=" + self.survey + " AND event_id=" + str(self.activeEvent) +
+                          " AND partition='MainTrawl' AND event_parameter='" + event_parameter + "'")
+        else:
+            exists_sql = ("SELECT gear_accessory_option FROM " + self.schema + ".gear_accessory WHERE ship="
+                          + self.ship + " AND survey=" + self.survey + " AND event_id=" + str(self.activeEvent) +
+                          " AND gear_accessory='" + event_parameter + "'")
         exists_query = self.db.dbQuery(exists_sql)
         exists, = exists_query.first()
         if exists:
@@ -308,7 +375,11 @@ class MetaDlg(QDialog, ui_MetaDlg.Ui_metaDlg):
             return 0
 
     def cancel(self):
-        self.close()
+        """
+        if the cancel button is pressed, reject the dialog to close
+        :return:
+        """
+        self.reject()
 
     def closeEvent(self, event=None):
         self.accept()
